+++ src/sdc11073/provider/scopesfactory.py
@@ -1,9 +1,34 @@
 """The module implements the function mk_scopes."""
-from urllib.parse import quote_plus
 
-from sdc11073.location import SdcLocation
+import urllib.parse
+
 from sdc11073.mdib.mdibprotocol import ProviderMdibProtocol
 from sdc11073.xml_types.wsd_types import ScopesType
+
+# from IEEE Std 11073-20701-2018 chapter 9.3 SDC PARTICIPANT KEY PURPOSE based discovery
+KEY_PURPOSE_SERVICE_PROVIDER = 'sdc.mds.pkp:1.2.840.10004.20701.1.1'
+# from IEEE Std 11073-20701-2018 chapter 9.4 Context-based discovery
+BICEPS_URI_UNK = 'biceps.uri.unk'
+
+
+def _query_from_location_state(state) -> str:  # noqa: ANN001  typing is unknown here as it depends on the data_model in the mdib
+    """Return a query string from a LocationContextStateContainer from GLUE 9.4.1.2."""
+    query_dict: dict[str, str] = {}
+    if state.LocationDetail is None:
+        return ''
+    if state.LocationDetail.Facility is not None:
+        query_dict['fac'] = state.LocationDetail.Facility
+    if state.LocationDetail.Building is not None:
+        query_dict['bldng'] = state.LocationDetail.Building
+    if state.LocationDetail.Floor is not None:
+        query_dict['flr'] = state.LocationDetail.Floor
+    if state.LocationDetail.PoC is not None:
+        query_dict['poc'] = state.LocationDetail.PoC
+    if state.LocationDetail.Room is not None:
+        query_dict['rm'] = state.LocationDetail.Room
+    if state.LocationDetail.Bed is not None:
+        query_dict['bed'] = state.LocationDetail.Bed
+    return urllib.parse.urlencode(query_dict, quote_via=urllib.parse.quote, safe='')
 
 
 def mk_scopes(mdib: ProviderMdibProtocol) -> ScopesType:
@@ -14,42 +39,50 @@ def mk_scopes(mdib: ProviderMdibProtocol) -> ScopesType:
     pm_types = mdib.data_model.pm_types
     pm_names = mdib.data_model.pm_names
     scope = ScopesType()
-    loc_entities = mdib.entities.by_node_type(pm_names.LocationContextDescriptor)
-    assoc_loc = []
-    for ent in loc_entities:
-        assoc_loc.extend([loc for loc in ent.states.values() if
-                          loc.ContextAssociation == pm_types.ContextAssociation.ASSOCIATED])
-    if len(assoc_loc) == 1:
-        loc = assoc_loc[0]
-        det = loc.LocationDetail
-        dr_loc = SdcLocation(fac=det.Facility, poc=det.PoC, bed=det.Bed, bldng=det.Building,
-                             flr=det.Floor, rm=det.Room)
-        scope.text.append(dr_loc.scope_string)
-
-    for nodetype, scheme in ((pm_names.OperatorContextDescriptor, 'sdc.ctxt.opr'),
-                             (pm_names.EnsembleContextDescriptor, 'sdc.ctxt.ens'),
-                             (pm_names.WorkflowContextDescriptor, 'sdc.ctxt.wfl'),
-                             (pm_names.MeansContextDescriptor, 'sdc.ctxt.mns')):
+    for nodetype, scheme in (
+        (pm_names.LocationContextDescriptor, 'sdc.ctxt.loc'),
+        (pm_names.OperatorContextDescriptor, 'sdc.ctxt.opr'),
+        (pm_names.EnsembleContextDescriptor, 'sdc.ctxt.ens'),
+        (pm_names.WorkflowContextDescriptor, 'sdc.ctxt.wfl'),
+        (pm_names.MeansContextDescriptor, 'sdc.ctxt.mns'),
+    ):
         entities = mdib.entities.by_node_type(nodetype)
         for entity in entities:
-            states = entity.states
-            assoc_st = [s for s in states if s.ContextAssociation == pm_types.ContextAssociation.ASSOCIATED]
-            for state in assoc_st:
+            for state in [
+                s for s in entity.states.values() if s.ContextAssociation == pm_types.ContextAssociation.ASSOCIATED
+            ]:
+                if not state.Identification:
+                    msg = f'State {state.Handle} of type {nodetype} has no Identification element'
+                    raise ValueError(msg)
                 for ident in state.Identification:
-                    scope.text.append(f'{scheme}:/{quote_plus(ident.Root)}/{quote_plus(ident.Extension)}')
+                    # IEEE Std 11073-20701-2018 9.4 context based discovery
+                    instance_identifier = (
+                        f'/{urllib.parse.quote(ident.Root if ident.Root is not None else BICEPS_URI_UNK, safe="")}'
+                    )
+                    if ident.Extension:
+                        instance_identifier += f'/{urllib.parse.quote(ident.Extension, safe="")}'
+                    context_uri = f'{scheme}:{instance_identifier}'
+                    query = ''
+                    if nodetype == pm_names.LocationContextDescriptor:
+                        if not state.LocationDetail:
+                            msg = f'State {state.Handle} of type {nodetype} has no LocationDetail element'
+                            raise ValueError(msg)
+                        query = _query_from_location_state(state)
+                    if query:
+                        context_uri = f'{context_uri}?{query}'
+                    scope.text.append(context_uri)
 
     scope.text.extend(_get_device_component_based_scopes(mdib))
-    scope.text.append('sdc.mds.pkp:1.2.840.10004.20701.1.1')  # key purpose Service provider
+    scope.text.append(KEY_PURPOSE_SERVICE_PROVIDER)  # default scope that is always included
     return scope
+
 
 def _get_device_component_based_scopes(mdib: ProviderMdibProtocol) -> set[str]:
     """Return a set of scope strings.
 
-    SDC: For every instance derived from pm:AbstractComplexDeviceComponentDescriptor in the MDIB an
-    SDC SERVICE PROVIDER SHOULD include a URI-encoded pm:AbstractComplexDeviceComponentDescriptor/pm:Type
-    as dpws:Scope of the MDPWS discovery messages. The URI encoding conforms to the given Extended Backus-Naur Form.
-    E.G.  sdc.cdc.type:///69650, sdc.cdc.type:/urn:oid:1.3.6.1.4.1.3592.2.1.1.0//DN_VMD
-    After discussion with David: use only MDSDescriptor, VmdDescriptor makes no sense.
+    Use only MDSDescriptor, because there might be alot of VmdDescriptor that might exceed the dpws message size limit.
+    Also, VmdDescriptor do not contain relevant information for discovery purposes.
+
     :return: a set of scope strings
     """
     pm_types = mdib.data_model.pm_types
@@ -58,8 +91,11 @@ def _get_device_component_based_scopes(mdib: ProviderMdibProtocol) -> set[str]:
     entities = mdib.entities.by_node_type(pm_names.MdsDescriptor)
     for entity in entities:
         if entity.descriptor.Type is not None:
-            coding_systems = '' if entity.descriptor.Type.CodingSystem == pm_types.DEFAULT_CODING_SYSTEM \
+            coding_systems = (
+                ''
+                if entity.descriptor.Type.CodingSystem == pm_types.DEFAULT_CODING_SYSTEM
                 else entity.descriptor.Type.CodingSystem
+            )
             csv = entity.descriptor.Type.CodingSystemVersion or ''
             scope_string = f'sdc.cdc.type:/{coding_systems}/{csv}/{entity.descriptor.Type.Code}'
             scopes.add(scope_string)