+++ src/sdc11073/wsdiscovery/common.py
@@ -1,3 +1,5 @@
+"""Common definitions for WS-Discovery module."""
+
 import logging
 
 from sdc11073.definitions_sdc import SdcV1Definitions
@@ -9,4 +11,3 @@ message_reader = MessageReader(SdcV1Definitions, None, logger=logging.getLogger(
 
 MULTICAST_PORT = 3702
 MULTICAST_IPV4_ADDRESS = "239.255.255.250"
-MULTICAST_OUT_TTL = 15  # Time To Live for multicast_out